# swagger/swagger.yml
info:
  title: Comentario
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
swagger: "2.0"
basePath: /api
securityDefinitions:
  # Cookie authentication for commenters
  commenterTokenHeader:
    type: apiKey
    in: header
    name: X-Commenter-Token

  # Cookie authentication for owners. Uses the apiKey type for the lack of a proper cookie-based authentication in
  # Swagger 2
  ownerCookie:
    type: apiKey
    in: header
    name: Cookie
...
parameters:
  federatedIdpId:
    in: path
    name: provider
    required: true
    description: Federated identity provider ID
    type: string
    enum:
      - github
      - gitlab
      - google
      - twitter
responses:
...
  OAuthRedirect:
    description: Redirecting to a federated identity provider
    headers:
      Location:
        type: string

  # 401
  OAuthFailure:
    description: Authentication failed
    schema:
      type: string
paths:
...
  /oauth/{provider}:
    get:
      operationId: OauthInit
      summary: Initiate signing in using federated authentication
      produces:
        - text/html # For an error response only, otherwise a redirect is issued
      parameters:
        - $ref: "#/parameters/federatedIdpId"
        - in: query
          name: commenterToken
          type: string
          required: true
          minLength: 64
          maxLength: 64
          pattern: '[0-9a-f]{64}'
      responses:
        307:
          $ref: "#/responses/OAuthRedirect"
        401:
          $ref: "#/responses/OAuthFailure"
  /oauth/{provider}/callback:
    get:
      operationId: OauthCallback
      summary: Endpoint that gets called back by the federated authentication provider
      produces:
        - text/html
      parameters:
        - $ref: "#/parameters/federatedIdpId"
      responses:
        200:
          description: User has successfully signed in
        401:
          $ref: "#/responses/OAuthFailure"
  /oauth/sso:
    get:
      operationId: OauthSsoInit
      summary: Initiate signing in using federated SSO authentication
      produces:
        - text/html # For an error response only, otherwise a redirect is issued
      parameters:
        - in: query
          name: commenterToken
          type: string
          required: true
          minLength: 64
          maxLength: 64
          pattern: '[0-9a-f]{64}'
      responses:
        307:
          $ref: "#/responses/OAuthRedirect"
        401:
          $ref: "#/responses/OAuthFailure"
  /oauth/sso/callback:
    get:
      operationId: OauthSsoCallback
      summary: Endpoint that gets called back by the SSO authentication provider
      produces:
        - text/html
      parameters:
        - in: query
          name: payload
          type: string
          required: true
          minLength: 1
        - in: query
          name: hmac
          type: string
          required: true
          minLength: 1
      responses:
        200:
          description: User has successfully signed in
        401:
          $ref: "#/responses/OAuthFailure"

  #---------------------------------------------------------------------------------------------------------------------
  # Testing endpoints
  # These endpoints are only available during the test cycle and never in production
  #---------------------------------------------------------------------------------------------------------------------
...